requirements.txt
Flask==3.0.2
markdown2==2.4.12
weasyprint==60.1
python-docx==1.1.0
markdown==3.5.2
pydyf==0.8.0
beautifulsoup4==4.12.3
pytest==8.0.0
requests==2.31.0
PyPDF2==3.0.1
